Refactor on main.js logic
