+++ src/index.js
@@ -1,50 +1,9 @@
-//Page to
-let flag = 0;
+import { addNavbar, changeTab } from "./navbar";
+
 function bodyComponent() {
   const element = document.createElement("div");
   element.className = "main-container";
   return element;
-}
-
-function addNavbar() {
-  const navContainer = document.createElement("div");
-  navContainer.className = "tabscontainer";
-
-  const nav = document.querySelector(".main-container");
-  nav.appendChild(navContainer);
-
-  const tabs = ["Home", "Menu", "Contact"];
-  tabs.forEach(tab => {
-    const button = document.createElement("button");
-    button.addEventListener("click", changeTab);
-    button.textContent = tab;
-    button.classList.add(tab.toLowerCase());
-
-    navContainer.appendChild(button);
-  });
-}
-
-function changeTab() {
-  const contact = document.getElementById("contact-container");
-  const menu = document.getElementById("menu-container");
-  const home = document.querySelector(".textHome");
-  if (this.className === "home") {
-    contact.style.display = "none";
-    menu.style.display = "none";
-    home.style.display = "inherit";
-  } else if (this.className === "menu") {
-    home.style.display = "none";
-    contact.style.display = "none";
-    menu.style.display = "inherit";
-  } else if (this.className === "contact") {
-    home.style.display = "none";
-    contact.style.display = "inherit";
-    menu.style.display = "none";
-  } else {
-    contact.style.display = "none";
-    menu.style.display = "none";
-    home.style.display = "inherit";
-  }
 }
 
 function showHome() {
@@ -107,8 +66,10 @@ function showMenu() {
 }
 
 document.body.appendChild(bodyComponent());
-addNavbar();
-showHome();
-showContact();
-showMenu();
-changeTab();
+(function() {
+  addNavbar();
+  showHome();
+  showContact();
+  showMenu();
+  changeTab();
+})();
